+++ tests/test_node.py
@@ -4,9 +4,9 @@ class TestNodes(object):
     def linear_graph(self):
         # Constructs a linear graph:
         # A -> B -> C
-        A = Node(name="A")
-        B = Node(inputs=[A], name="B")
-        C = Node(inputs=[B], name="C")
+        A = Node(path="A")
+        B = Node(inputs=[A], path="B")
+        C = Node(inputs=[B], path="C")
         return A, B, C
 
     def test_linear_outputs_correct(self):
@@ -21,10 +21,10 @@ class TestNodes(object):
         # B     C
         #   \  /
         #    D (output)
-        A = Node(name="A")
-        B = Node(inputs=[A], name="B")
-        C = Node(inputs=[A], name="C")
-        D = Node(inputs=[B, C], name="D")
+        A = Node(path="A")
+        B = Node(inputs=[A], path="B")
+        C = Node(inputs=[A], path="C")
+        D = Node(inputs=[B, C], path="D")
         return A, B, C, D
 
     def test_diamond_outputs_correct(self):
@@ -32,3 +32,17 @@ class TestNodes(object):
         assert B in A.outputs and C in A.outputs
         assert D in B.outputs
         assert D in C.outputs
+
+    def test_hashes_same_node(self):
+        A0 = Node(path="A")
+        A1 = Node(path="A")
+        assert hash(A0) == hash(A1)
+        test = set([A0, A1])
+        assert len(test) == 1
+
+    def test_hashes_different_node(self):
+        A = Node(path="A")
+        B = Node(path="B")
+        assert hash(A) != hash(B)
+        test = set([A, B])
+        assert len(test) == 2